+++ CHANGELOG.md
@@ -25,6 +25,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.7.1] - 2021-07-27
+
+### Added
+- Added ``--excluded-ncns`` option to ``sat bootsys`` that can be used to omit
+  NCNs from the platform-services and ncn-power stages in case they are
+  inaccessible.
+
 ## [3.7.0] - 2021-05-13
 
 ### Changed